+++ tests/testthat/test_interaction.R
@@ -25,9 +25,6 @@ test_that("matched samples interactions work", {
 
   expect_true(mcols(y)$pvalue[2] < .01)
 
-  # try the old, slower method
-  y <- swish(y, x="condition", cov="group", pair="pair", interaction=TRUE, fast=0, quiet=TRUE)
-
 })
 
 test_that("two group interactions work", {